refactor: 통계 관련 메서드를 StatisticsController로 분리
- LottoController에 있던 통계 관련 메서드를 StatisticsController 클래스로 이동
- 코드의 모듈화를 통해 각 클래스의 역할을 명확히 구분
