+++ src/main/java/lotto/LottoController.java
@@ -35,20 +35,4 @@ public class LottoController {
             throw new IllegalArgumentException("[ERROR] 올바른 로또 번호 형식이 아닙니다.");
         }
     }
-
-    public LottoResult calculateWinnings(List<Lotto> lottos, WinNumbers winNumbers) {
-        int totalWinnings = 0;
-        int[] rankCount = new int[Rank.values().length];
-        for (Lotto lotto : lottos) {
-            Rank rank = lotto.matchWinNumbers(winNumbers.getWinNumbers(), winNumbers.getBonusNumber());
-            rankCount[rank.ordinal()]++;
-            totalWinnings += rank.getPrizeMoney();
-        }
-        return new LottoResult(totalWinnings, rankCount);
-    }
-
-    public double calculateRateOfReturn(int totalWinnings, int purchaseNumber) {
-        double rateOfReturn = ((double) totalWinnings / purchaseNumber) * 100;
-        return Math.round(rateOfReturn) / 10.0;
-    }
 }
